% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wobbiesworld.R
\name{learn}
\alias{learn}
\title{learn}
\usage{
learn(n = 10000, maze = makeWobbiesWorld(),
  control = makeRandomController(maze), verbose = TRUE, doPlot = TRUE,
  every = 100, col = 1, ymin = -1000)
}
\arguments{
\item{n}{Number of runs/games to perform}

\item{maze}{The maze to use. You should keep the default (with a single monster), which is
what will be used for evaluation.}

\item{control}{The control system. See runGame for details about this object's required characteristics.}

\item{verbose}{Should additional output be printed in the console.}

\item{doPlot}{Should a plot be drawn showing the evolution in performance of the system being trained}

\item{every}{If doPlot is TRUE, a new point is added to the plot after this many runs/games, showing the
average performance by the control system over the games since the last point was plotted,}

\item{col}{If plot is TRUE, the color of the point plotted}

\item{ymin}{If plot is TRUE, the lower limit of the area plotted (change to zoom in on good performance, rather
than have the plot dominated by early terrible performance).}
}
\value{
A list containing (1) The updated controller, (2) The vector of scores, (3) The vector of turns taken
}
\description{
The learn function used by the runWW function. It is exported in case you wish to call it directly.
}
