% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wobbiesworld.R
\name{makeRandomController}
\alias{makeRandomController}
\title{makeRandomController}
\usage{
makeRandomController(maze)
}
\arguments{
\item{The}{maze used}
}
\value{
A random controller able to be used in runWW and runGame.
}
\description{
This function returns a demo controller that just makes random moves (even when
doRand is set to FALSE).
}
