% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maze.R
\name{makeWobbiesWorld}
\alias{makeWobbiesWorld}
\title{Make the maze}
\usage{
makeWobbiesWorld(monsters = 1)
}
\arguments{
\item{monsters}{How many monsters to place in the maze. Default is 1, and you will be evaluated with 1.}
}
\value{
A list giving the maze object with the fields as described above.
}
\description{
This function makes the maze, and you can call this to familiarize yourself with the maze
object. You will use this object to set up your controller, and to recieve
information from the environment in every turn. The fields of the maze list are given in details below.
}
\details{
df - A data frame of possible x and y positions (this gives the maze)

wobbie - The location of Wobbie

goal - The location of the goal

monster1 - The location of the monster

reward - Reward from the last turn

alive - If Wobbie is alive

finished - If the game is over (won)

monsters - How many monsters are in the maze

lastAction - The last action taken by Wobbie
}
