% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wobbiesworld.R
\name{runGame}
\alias{runGame}
\title{runGame}
\usage{
runGame(maze = makeWobbiesWorld(), control = makeRandomController(maze),
  seed = NA, doPlot = F, pause = 1)
}
\arguments{
\item{maze}{The maze to use. You should keep the default (with a single monster), which is
what will be used for evaluation.}

\item{control}{The control system. See above for details.}

\item{seed}{The seed to use for this game. Ignored if NA}

\item{doPlot}{Should the game be plotted}

\item{pause}{If the game is plotted, the pause between moves. (Non-zero pause required for
plotting to work on most systems.)}
}
\value{
A list containing (1) The control system (updated if it is learning), (2) The game score, (3) The turns taken.
}
\description{
Runs a single game of Wobbie's World.
}
\details{
You need to pass a controller to this function.This controller should be a list, containing two named fields:

decideAction: A function which takes as arguments the controller and
      the maze, and returns an action (2=down,4=left,6=right,8=up)

update: A function which takes as arguments the controller and the maze,
      and returns the (updated) controller.

Each turn the decideAction function is called. The maze is then updated
based on the returned action, and then the update function is called with
the updated maze so the controller can learn the consequences of its
action.

The game terminates when Wobbie exits the maze or when he is caught by the monster.
The game score is the number of turns taken, minus 1000 if Wobbie was caught by the monster.
}
