% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wobbiesworld.R
\name{runWW}
\alias{runWW}
\title{runWW}
\usage{
runWW(makeController = makeRandomController, verbose = FALSE,
  doPlot = FALSE, seed = NA)
}
\arguments{
\item{makeController}{A function that will return a Q-learning controller. See above for
details.}

\item{verbose}{Passed to the learn function}

\item{doPlot}{Passed to the learn function}

\item{seed}{A random seed to be used before anything else is executed.}
}
\value{
A vector containing (1) the mean performance of the test runs; and (2) the time taken in seconds.
}
\description{
You will need to pass a function that creates a Q-learning RL controller.
This controller will then play 10000 games, learning to control
the Wobbie's World game. It will be tested on 500 games after it
is finished learning. This function should take the maze object
returned from the makeWobbiesWorld function as the only argument. (The idea
is you can use the maze object to help set up your controller.)
}
\details{
Once created your controller will be passed to the learn function.

Your controller should have a field called doRand. This will be
set to FALSE after learning is completed, and you should set things
up so that your controller will no longer make random moves when this
is FALSE.

Once learning is complete, your controller will be passed to the test
function and performance over 500 test runs will be recorded.

Assessment will be based on the performance of your function for
a particular seed. It is compared to a par table-based q-learning
controller, and your average over the 500 test games need to be
at least within 1 of the par. For seed 0, the par controller
got an average score of 985.74 with a standard deviation of approximately
6.55.

On my laptop, the par controller took approximately 56 seconds to complete
(with verbose and doPlot both FALSE). To pass your code will need to complete
no more than 25% slower than the par function, or 300 seconds, whichever is higher.
}
